metadata-service/pyproject.toml
[project]
name = "python-dev-environment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "pydantic-settings>=2.7.1",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.3.4",
    "ruff>=0.9.1",
]
